pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image_similarity_search"
version = "0.0.0"
authors = [{ name = "chuzhdontcode" }]
description = "Image similarity search engine using FAISS with CLIP, SSCD, and DINOv2 support"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "faiss-gpu",                # or faiss-gpu for CUDA support
    "transformers>=4.30.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0,<2.0.0",
    "tqdm>=4.65.0",
    "pytorch-lightning>=2.0.0",
    "pandas>=2.0.0",
    "huggingface-hub>=0.16.0",
]

[project.optional-dependencies]
gpu = ["faiss-gpu"]
cpu = ["faiss-cpu>=1.7.4"]

[tool.setuptools]
packages = ["image_similarity_search"]
package-dir = { "" = "." }

[tool.setuptools.package-data]
image_similarity_search = ["*"]
